+++ src/pais/__init__.py
@@ -24,4 +24,4 @@ __all__ = [
     "Settings",
 ]
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"